# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.CategoryInput:
    properties:
      name:
        example: Alimentação
        type: string
      parent_id:
        description: UUID da categoria pai (opcional)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      sequence:
        example: 1
        minimum: 1
        type: integer
    type: object
  dtos.CategoryOutput:
    properties:
      children:
        items:
          $ref: '#/definitions/dtos.CategoryOutput'
        type: array
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Alimentação
        type: string
      sequence:
        example: 1
        type: integer
    type: object
  dtos.CategoryPaginatedOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.CategoryOutput'
        type: array
      pagination:
        $ref: '#/definitions/dtos.CategoryPaginationMeta'
    type: object
  dtos.CategoryPaginationMeta:
    properties:
      has_next:
        example: false
        type: boolean
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJmIjp7Im5hbWUiOiJBbGltZW50YcOnw6NvIn19
        type: string
    type: object
...
  httperrors.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        additionalProperties: {}
        type: object
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Jailton Junior
    url: https://github.com/jailtonjunior94
  description: |-
    API para gestão financeira pessoal com suporte a cartões de crédito, transações,
    orçamentos mensais, categorias e faturas com parcelamento.

    ## Valores Monetários
    Todos os valores monetários são representados como **strings decimais** (ex: `"1234.56"`)
    para preservar precisão e evitar erros de ponto flutuante.

    ## Paginação
    Endpoints de listagem usam **cursor-based pagination**. Parâmetros:
    - `limit` (int, default: 20, max: 100)
    - `cursor` (string, opcional): valor de `pagination.next_cursor` da resposta anterior

    ## Autenticação
    Endpoints protegidos requerem o header `Authorization: Bearer {token}`.
    Obtenha o token em `POST /api/v1/token`.
  license:
    name: MIT
  title: Financial API
  version: 1.0.0
paths:
...
  /api/v1/categories:
    get:
      description: Retorna a lista paginada de categorias do usuário autenticado (cursor-based
        pagination).
      parameters:
      - default: 20
        description: 'Itens por página (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Cursor de paginação
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de categorias
          schema:
            $ref: '#/definitions/dtos.CategoryPaginatedOutput'
        "400":
          description: Parâmetro inválido
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar categorias
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: |-
        Cria uma nova categoria para o usuário autenticado.
        Categorias podem ser hierárquicas: informe `parent_id` para criar uma subcategoria.
        - `parent_id`: UUID da categoria pai (opcional)
        - `sequence`: ordem de exibição (mínimo 1)
      parameters:
      - description: Dados da categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Categoria criada com sucesso
          schema:
            $ref: '#/definitions/dtos.CategoryOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "409":
          description: Ciclo detectado na hierarquia
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Criar categoria
      tags:
      - categories
  /api/v1/categories/{id}:
    delete:
      description: Remove uma categoria do usuário autenticado. Esta operação é irreversível.
      parameters:
      - description: ID da categoria
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Categoria removida com sucesso
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Categoria não encontrada
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remover categoria
      tags:
      - categories
    get:
      description: Retorna os detalhes de uma categoria específica, incluindo subcategorias
        (`children`).
      parameters:
      - description: ID da categoria
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados da categoria com filhos
          schema:
            $ref: '#/definitions/dtos.CategoryOutput'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Categoria não encontrada
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Buscar categoria por ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma categoria existente do usuário autenticado.
      parameters:
      - description: ID da categoria
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados da categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Categoria atualizada
          schema:
            $ref: '#/definitions/dtos.CategoryOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Categoria não encontrada
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "409":
          description: Ciclo detectado na hierarquia
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Atualizar categoria
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: `Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"